Moved files to designated components folder
