+++ src/app/(landing)/components/hero.tsx
