+++ src/app/(landing)/components/navbar.tsx
@@ -9,7 +9,7 @@ import { cn } from "@/lib/utils";
 import ScrollLink from "@/components/ui/scroll-link";
 import { Button, buttonVariants } from "@/components/ui/button";
 import ThemeToggle from "@/components/ui/ThemeToggle";
-import MobileNav from "@/components/home/MobileNav";
+import MobileNav from "@/app/(landing)/components/mobile-nav";
 import { homeRoutes } from "@/lib/routes";
 
 export default function Navbar() {
