+++ src/app/(landing)/components/section-about.tsx
