+++ src/app/(landing)/components/section-message.tsx
