+++ src/app/(landing)/components/section-programs.tsx
