+++ src/components/navbar/faculty-navbar.tsx
@@ -1,10 +1,10 @@
-import { MobileAdminSide } from "@/components/sidebar/MobileAdminSide";
+import { MobileFacultySide } from "@/components/sidebar/mobile-faculty-sidebar";
 import UserAccountDrop from "@/components/user-dropdown";
 
-export default function AdminNavbar() {
+export default function FacultyNavbar() {
   return (
     <div className="flex items-center p-4">
-      <MobileAdminSide />
+      <MobileFacultySide />
       <div className="flex items-center w-full justify-end">
         <UserAccountDrop />
       </div>
