+++ tsconfig.json
@@ -24,7 +24,23 @@
       "@/styles/*": ["./src/styles/*"],
       "@/types/*": ["./src/types/*"],
       "@/lib/*": ["./src/lib/*"],
-      "@/components/*": ["./src/components/*"],
+      "@/components/*": [
+        "./src/components/*", 
+        "./src/app/(landing)/components/*", 
+        "./src/app/(admin)/admin/(dashboard)/components/*",
+        "./src/app/(admin)/admin/enrollment/components/*", 
+        "./src/app/(admin)/admin/manage-accounts/components/*",
+        "./src/app/(admin)/admin/request/components/*",
+        "./src/app/(admin)/admin/student-record/components/*",
+        "./src/app/(auth)/components/*",
+        "./src/app/(faculty)/faculty/(dashboard)/components/*",
+        "./src/app/(faculty)/faculty/advisory-class/components/*",
+        "./src/app/(faculty)/faculty/grades/components/*",
+        "./src/app/(student)/student/(dashboard)/components/*",
+        "./src/app/(student)/student/enrollment/components/*",
+        "./src/app/(student)/student/grades/components/*",
+        "./src/app/(student)/student/request/components/*",
+      ],
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
